+++ packages/core/src/agent/polling/index.ts
@@ -24,6 +24,7 @@ import type { PollStrategy, PollForResponseResult } from './types.ts';
 import { ReadRequestType, type ReadStateRequest } from '../agent/http/types.ts';
 import { RequestStatusResponseStatus } from '../agent/http/index.ts';
 import { utf8ToBytes } from '@noble/hashes/utils';
+import type { Identity } from '../auth.ts';
 export { defaultStrategy } from './strategy.ts';
 export type { PollStrategy, PollForResponseResult } from './types.ts';
 
@@ -118,6 +119,7 @@ function isSignedReadStateRequestWithExpiry(
  * @param canisterId The effective canister ID.
  * @param requestId The Request ID to poll status for.
  * @param options polling options to control behavior
+ * @param identity - (Optional) The identity to use for the polling requests. If not provided, the agent's current identity will be used.
  * @returns The certificate, reply bytes, and raw certificate bytes for the request.
  * @throws {ExternalError} If the agent's root key is not available.
  * @throws {RejectError} If the request was rejected by the canister.
@@ -128,6 +130,7 @@ export async function pollForResponse(
   canisterId: Principal,
   requestId: RequestId,
   options: PollingOptions = {},
+  identity?: Identity | Promise<Identity>,
 ): Promise<PollForResponseResult> {
   const path = [utf8ToBytes('request_status'), requestId];
 
@@ -141,10 +144,10 @@ export async function pollForResponse(
       agent,
       pollingOptions: options,
     });
-    state = await agent.readState(canisterId, { paths: [path] }, undefined, currentRequest);
+    state = await agent.readState(canisterId, { paths: [path] }, identity, currentRequest);
   } else {
     // If preSignReadStateRequest is false, we use the default strategy and sign the request each time
-    state = await agent.readState(canisterId, { paths: [path] });
+    state = await agent.readState(canisterId, { paths: [path] }, identity);
   }
 
   if (agent.rootKey == null) {
@@ -188,7 +191,7 @@ export async function pollForResponse(
         // Pass over either the strategy already provided or the new one created above
         strategy,
         request: currentRequest,
-      });
+      }, identity);
     }
 
     case RequestStatusResponseStatus.Rejected: {